requirements.txt
absl-py==2.3.1
alembic==1.16.5
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
beautifulsoup4==4.14.2
black==25.9.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
clearml==2.0.3rc0
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
crc32c==2.7.1
cryptography==45.0.7
cyclopts==3.24.0
databricks-sdk==0.67.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
donfig==0.8.1.post1
efficientnet_pytorch==0.7.1
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.118.2
fastmcp==2.12.4
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
fsspec==2025.10.0
furl==2.1.4
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.75.1
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
huggingface_hub==1.1.2
humanize==4.13.0
idna==3.4
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.1.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.12.0
lazy_loader==0.4
lmdb==1.7.3
lpips==0.1.4
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
ml_dtypes==0.5.3
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
monai==1.5.1
more-itertools==10.8.0
mpmath==1.3.0
munch==4.0.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
numcodecs==0.16.3
numpy==2.3.4
onnx==1.19.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.10.0
opencv-python-headless==4.10.0
openslide-bin==4.0.0.8
openslide-python==1.4.2
optuna==4.5.0
packaging==23.4
parse==1.20.2
parso==0.8.4
pathable==0.4.4
pathlib2==2.3.7.post1
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.3.7
pretrainedmodels==0.7.4
prompt_toolkit==3.0.50
protobuf==6.32.1
psutil==7.0.0
pyamg==5.2.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pydicom==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
pynrrd==1.1.3
pyperclip==1.11.0
PyQt6==6.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-ignite==0.4.11
PyYAML==6.0.3
pyzmq==26.4.0
requests==2.31.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
seaborn==0.12.2
segmentation-models-pytorch==0.3.2
setuptools==80.9.0
shellingham==1.5.4
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.10.4
torch==2.9.0
torchio==0.20.23
torchvision==0.24.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typer-slim==0.20.0
typing_extensions==4.15.0
ultralytics==8.3.223
ultralytics-thop==2.0.18
uvicorn==0.37.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.1
wheel==0.45.1
wrapt==1.17.3
zarr==3.1.3
zipp==3.23.0
zstandard==0.25.0
